fixed scroll feature and organized file structure
