+++ src/components/CardList.js
@@ -1,20 +1,17 @@
 import React from 'react';
-import robots from './Robots';
+import robots from '../constants/Robots';
 import Card from './Card';
 
 
 const CardList = ({robots}) => {
     return(
-        <>
-        {
-        
+    <>
+    {
         robots.map((user, i) => {return(
             <Card id={robots[i].id} email={robots[i].email} name={robots[i].name}/>
-        );
-    })
-}
-        </>
+        );})
+    }
+    </>
     );
-    
 };
 export default CardList;
